Images - Timestamp fixer
Fixes the timestamps of newly copied images with the original timestamps from source files.

Usecase: copying files from a position to another, thus writing a new file, will mess up the creation date. This script will fix that.
The name must be the same, the only thing that can vary is the format.